requirements.txt
click==8.1.7
Flask==1.1.2
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
python-dotenv==0.19.2
Werkzeug==3.0.3
